dom/html/test/dialog/test_dialog_keydown_preventDefault.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1322947
-->
<head>
  <title>Test for Bug 1322947</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="SimpleTest.waitForFocus(runTest)">
  <a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=1322947">Test cancel event with preventDefault on keydown event for dialog element</a>
<p id="display"></p>
<dialog>
  <p>Hello World</p>
</dialog>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

var hasCancelEventFired = false;

function runTest() {
  const dialog = document.querySelector("dialog");

  const verify = () => {
    ok(!hasCancelEventFired, "cancel should not be fired");
    ok(hasKeydownEventFired, "document level keydown event should be fired");
    done();
  }

  dialog.addEventListener("cancel", function(event) {
    hasCancelEventFired = true;
  });

  document.addEventListener("keydown", function(event) {
    hasKeydownEventFired = true;
    event.preventDefault();
    setTimeout(function() {
      verify();
    }, 0);
  });
  dialog.showModal();
  synthesizeKey("VK_ESCAPE", {}, window);
}

function done() {
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
